refactor:#128-edit NotFoundPage.js in upper case
